pyproject.toml
[project]
name = "taskmaster-pipefocus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "fastapi>=0.128.0",
    "google-api-python-client>=2.188.0",
    "google-auth>=2.48.0",
    "google-auth-httplib2>=0.3.0",
    "google-auth-oauthlib>=1.2.4",
    "jinja2>=3.1.6",
    "pandas>=2.3.3",
    "plotly>=6.5.2",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "requests>=2.32.5",
    "sqlmodel>=0.0.32",
    "streamlit>=1.53.1",
    "uvicorn>=0.40.0",
]

[dependency-groups]
dev = [
    "ruff>=0.14.14",
]

[tool.ruff]
line-length = 88
target-version = "py312" # Ou sua versão instalada

[tool.ruff.lint]
select = ["E", "F", "I"] # Style errors, Flakes, Isort (Import sorting)

[tool.ruff.format]
quote-style = "double"
indent-style = "tab"
